% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{get_op3_token}
\alias{get_op3_token}
\title{Obtain OP3 API token}
\usage{
get_op3_token()
}
\value{
list with API token
}
\description{
Obtain the OP3 API token from a \code{.Renviron} file, either stored in the
current directory or the user's home directory.
}
