% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{op3_downloads_show}
\alias{op3_downloads_show}
\title{Get monthly and weekly show download counts}
\usage{
op3_downloads_show(show_id, nest_downloads = TRUE)
}
\arguments{
\item{show_id}{One or more character strings of OP3 show UUID values}

\item{nest_downloads}{Boolean to create a list column of the weekly
download metrics or expand the data frame to have one row per
week of download metrics. Default is \code{TRUE}.}
}
\value{
\code{tibble} data frame with the following columns:
\itemize{
\item \code{days}: Number of days in time window
\item \code{monthlyDownloads}: Number of downloads in month
\item \code{weeklyAvgDownloads}: Average number of downloads per week in month
\item \code{numWeeks}: Number of weeks in time window
\item \code{download_data}: Nested data frame of download numbers per week (when \code{nest_downloads} is TRUE)
\item \code{weekNumber}: Week number in time window (when \code{nest_downloads} is FALSE)
\item \code{weeklyDownloads}: Number of downloads for \code{weekNumber} (when \code{nest_downloads} is FALSE)
}
}
\description{
\code{op3_downloads_show()} obtains the number of monthly downloads and average
weekly downloads over the last four weeks (excludes bots)
}
\examples{
\dontshow{if (op3r::op3_token_isset()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Requires API token

op3_downloads_show(show_id = "a18389b8a52d4112a782b32f40f73df6")
\dontshow{\}) # examplesIf}
}
