% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shows.R
\name{op3_get_show_uuid}
\alias{op3_get_show_uuid}
\title{Get podcast OP3 UUID from Podcast GUID or RSS Feed}
\usage{
op3_get_show_uuid(show_guid = NULL, show_rss_url = NULL)
}
\arguments{
\item{show_guid}{String representing the podcast GUID.}

\item{show_rss_url}{String representing the podcast RSS feed}
}
\value{
string with the podcast UUID
}
\description{
\code{op3_get_show_uuid()} uses the API to obtain the podcast
OP3 UUID based on either the podcast GUID or podcast RSS feed.
Only one of the two can be specified.
}
