% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hits.R
\name{op3_hits}
\alias{op3_hits}
\title{Get log record metadata for OP3 service}
\usage{
op3_hits(
  format = "json",
  limit = 100,
  desc = TRUE,
  start = NULL,
  start_inclusive = TRUE,
  end = NULL,
  url = NULL,
  hashed_ip_address = NULL
)
}
\arguments{
\item{format}{result output format returned by the API endpoint. Possible
values are \code{json} (object-based) or \code{json-a} (array-based) formats.}

\item{limit}{Integer of maximum number of episodes to return. The
maximum limit supported for this endpoint is \code{1000}. Default is \code{100}.}

\item{desc}{Boolean to return results in time-descending order. If \code{FALSE},
results are returned in time-ascending order. Default is \code{TRUE}.}

\item{start}{Optional date or date-time object to filter results to those
starting at the supplied value. The object must be class \code{Date}. Default
is \code{NULL}.}

\item{start_inclusive}{Boolean to consider the \code{start} date or date-time
as inclusive for results. If \code{FALSE}, the \code{start} value is considered
exclusive. Default is \code{TRUE}.}

\item{end}{Optional date or date-time object to filter results to those
ending at the supplied value. The object must be class \code{Date}. Default
is \code{NULL}.}

\item{url}{Optional string to fulter results by the specific URL providing
the redirect. Default is \code{NULL}.}

\item{hashed_ip_address}{Optional string to filter results by the specific
IP address secure hash. Default is \code{NULL}.}
}
\value{
A\code{tibble} data frame if json format, otherwise a nested list if
json-a format.
}
\description{
\code{op3_hits()} obtains lowest-level log records for every prefix redirect
processed by the OP3 service.
}
\examples{
\dontshow{if (op3r::op3_token_isset()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Requires API token

op3_hits(limit = 10)
\dontshow{\}) # examplesIf}
}
