% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shows.R
\name{op3_show}
\alias{op3_show}
\title{Get podcast show information}
\usage{
op3_show(show_id, episodes = FALSE)
}
\arguments{
\item{show_id}{String representing an OP3 show UUID, podcast GUID, or
a podcast RSS feed URL.}

\item{episodes}{Boolean to return episode-level metadata along with
the show-level information. Default is \code{FALSE}.}
}
\value{
\code{tibble} data frame with the following columns:
\itemize{
\item \code{showUuid}: Podcast OP3 UUID
\item \code{title}:  Podcast title
\item \code{podcastGuid}: Podcast GUID
\item \code{statsPageUrl}: URL of the OP3 statistics web page for the podcast
\item \code{episodes_id}: Episode ID (when \code{episodes} is TRUE)
\item \code{episodes_title}: Episode title (when \code{episodes} is TRUE)
}
}
\description{
\code{op3_show()} obtains show-level information for a given podcast
using OP3.
}
\examples{
\dontshow{if (op3r::op3_token_isset()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Requires API token

op3_show(show_id = "bb28afcc-137e-5c66-b231-4ffad7979b44")
\dontshow{\}) # examplesIf}
}
