% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{op3_token_isset}
\alias{op3_token_isset}
\title{Detect if OP3 API token is defined}
\usage{
op3_token_isset()
}
\value{
boolean, TRUE if 'OP3_API_TOKEN` is defined in the current
R session.
}
\description{
Detect if OP3 API token is defined
}
