% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{op3_top_apps}
\alias{op3_top_apps}
\title{Get top apps by share of downloads}
\usage{
op3_top_apps(device_name = NULL)
}
\arguments{
\item{device_name}{Optional string of a specific device ID to constrain
results.}
}
\value{
\code{tibble} data frame with the following columns:
\itemize{
\item \code{app_name}: Device name
\item \code{value}: Percentage of download share
\item \code{min_date}: Oldest date in data range
\item \code{max_date}: Newest date in data range
}
}
\description{
\code{op3_top_apps()} obtains global shares of application downloads over the
last thirty days categorized by application
}
\examples{
\dontshow{if (op3r::op3_token_isset()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Requires API token

op3_top_apps()
\dontshow{\}) # examplesIf}
}
