% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{op3_top_show_apps}
\alias{op3_top_show_apps}
\title{Get download numbers by application for a show}
\usage{
op3_top_show_apps(show_id)
}
\arguments{
\item{show_id}{String representing an OP3 show UUID, podcast GUID, or
a podcast RSS feed URL.}
}
\value{
\code{tibble} data frame with the following columns:
\itemize{
\item \code{show_uuid}: OP3 show UUID
\item \code{app_name}: Name of podcast application
\item \code{value}: Number of downloads in the last three calendar months
}
}
\description{
\code{op3_top_show_apps()} obtains the download numbers for individual
podcast applications over the last three calendar months
for a specific show.
}
\examples{
\dontshow{if (op3r::op3_token_isset()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Requires API token

op3_top_show_apps(show_id = "bb28afcc-137e-5c66-b231-4ffad7979b44")
\dontshow{\}) # examplesIf}
}
