% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hits.R
\name{op3_hits}
\alias{op3_hits}
\title{Get log record metadata for OP3 service}
\usage{
op3_hits(
  format = "json",
  limit = 100,
  desc = TRUE,
  start = NULL,
  start_inclusive = TRUE,
  end = NULL,
  url = NULL,
  hashed_ip_address = NULL
)
}
\arguments{
\item{format}{result output format returned by the API endpoint. Possible
values are \code{json} (object-based) or \code{json-a} (array-based) formats.}

\item{limit}{Integer of maximum number of episodes to return. The
maximum limit supported for this endpoint is \code{1000}. Default is \code{100}.}

\item{desc}{Boolean to return results in time-descending order. If \code{FALSE},
results are returned in time-ascending order. Default is \code{TRUE}.}

\item{start}{Optional date or date-time object to filter results to those
starting at the supplied value. The object must be class \code{Date}. Default
is \code{NULL}.}

\item{start_inclusive}{Boolean to consider the \code{start} date or date-time
as inclusive for results. If \code{FALSE}, the \code{start} value is considered
exclusive. Default is \code{TRUE}.}

\item{end}{Optional date or date-time object to filter results to those
ending at the supplied value. The object must be class \code{Date}. Default
is \code{NULL}.}

\item{url}{Optional string to filter results by the specific URL providing
the redirect. The URL is often a direct link to a podcast episode file,
or a version of the URL with a wildcard character \code{*} to enable multiple
episodes for a given podcast. Default is \code{NULL}.}

\item{hashed_ip_address}{Optional string to filter results by the specific
IP address secure hash. Default is \code{NULL}.}
}
\value{
A\code{tibble} data frame for the \code{json} format, otherwise a nested list
for the \code{json-a} format. The data frame contains the following columns:
\itemize{
\item \code{time}: Timestamp of current record
\item \code{uuid}: OP3 UUID for podcast
\item \code{hashedIpAddress}: Secure hash of the record's originating IP address
\item \code{method}: Type of HTTP method used
\item \code{url}: URL of the record's request
\item \code{userAgent}: Request's reported user agent
\item \code{range}: If application, the request's specified bytes range.
\item \code{edgeColo}: Three-letter airport code associated with the record
\item \code{continent}: The request's originating continent
\item \code{country}: Two-letter country code of the record
\item \code{timezone}: Request timezone
\item \code{regionCode}: Abbreviation of region
\item \code{region}: Region of request
\item \code{metroCode}: Region metro code
}
}
\description{
\code{op3_hits()} obtains lowest-level log records for every prefix redirect
processed by the OP3 service.
}
\examples{
\dontshow{if (op3r::op3_token_isset()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Requires API token

# obtain 10 records
op3_hits(limit = 10)

# records based on specific podcast episode
op3_hits(url = "https://op3.dev/e/serve.podhome.fm/episode/99cfd30f-e40c-426a-3557-08dc4ea63bb0/63851340872787683054574074-0a1f-4951-8f2a-9084678c6604v1.mp3")

# records for multiple episodes using a wildcard in URL
op3_hits(url = "https://op3.dev/e/serve.podhome.fm/episode/99cfd30f-e40c-426a-3557-08dc4ea63bb0/*")
\dontshow{\}) # examplesIf}
}
