% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{op3_transcripts}
\alias{op3_transcripts}
\title{Get recent episodes with transcripts}
\usage{
op3_transcripts(limit = 100)
}
\arguments{
\item{limit}{Integer of maximum number of episodes to
return. The maximum limit supported is 100.}
}
\value{
\code{tibble} data frame with the following columns:
\itemize{
\item \code{asof}: Timestamp
\item \code{pubdate}: Episode publish timestamp
\item \code{podcastGuid}: Podcast GUID
\item \code{episodeItemGuid}: Episode GUID
\item \code{hasTranscripts}: Boolean indicating if the episode has a transcript
\item \code{dailyDownloads}: List of episode download numbers by date
}
}
\description{
\code{op3_transcripts()} obtains the most recent episodes of podcasts using OP3
containing transcripts along with their daily download numbers.
}
